fixed a bug in weather class

+++ app/src/main/java/edu/uiuc/cs427app/Weather.java
@@ -54,7 +54,7 @@ public class Weather {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
